refactoring and added dmsrc operation
